+++ drivers/video/tegra/host/nvmap.h
@@ -44,9 +44,9 @@ void *nvhost_nvmap_kmap(struct mem_handle *handle, unsigned int pagenum);
 void nvhost_nvmap_kunmap(struct mem_handle *handle, unsigned int pagenum,
 		void *addr);
 struct mem_handle *nvhost_nvmap_get(struct mem_mgr *mgr,
-		u32 id, struct platform_device *dev);
+		ulong id, struct platform_device *dev);
 
-phys_addr_t nvhost_nvmap_get_addr_from_id(u32 id);
+phys_addr_t nvhost_nvmap_get_addr_from_id(ulong id);
 
 int nvhost_nvmap_pin_array_ids(struct mem_mgr *mgr,
 		u32 *ids,
@@ -56,7 +56,7 @@ int nvhost_nvmap_pin_array_ids(struct mem_mgr *mgr,
 		struct nvhost_job_unpin *unpin_data,
 		dma_addr_t *phys_addr);
 
-void nvhost_nvmap_unpin_id(struct mem_mgr *mgr, u32 id);
+void nvhost_nvmap_unpin_id(struct mem_mgr *mgr, ulong id);
 
 
 #endif
